disabled kuma appending and fix for account deletion mail

+++ src/instance/deletion.rs
@@ -152,8 +152,10 @@ pub async fn check_instance_standing() -> bool {
             std::fs::remove_file(warning_sent_path).warn("Removing warning sent file");
         }
         InstanceStanding::AlmostDeleted => {
-            email::send_deletion_warning_mail().warn("sending deletion warning");
-            std::fs::write(warning_sent_path, []).warn("writing deletion sent warning");
+            if warning_sent_path.exists() {
+                email::send_deletion_warning_mail().warn("sending deletion warning");
+                std::fs::write(warning_sent_path, []).warn("writing deletion sent warning");
+            }
         }
         InstanceStanding::MustDelete => {
             delete_account(user.id, DeletedReason::OldAge)
